refactor(api): make the seed scripts clearer
