+++ apps/api/package.json
@@ -7,7 +7,7 @@
     "node": ">=16.14.0"
   },
   "prisma": {
-    "seed": "node prisma/seed.js"
+    "seed": "node prisma/seed-development.js"
   },
   "type": "module",
   "scripts": {
@@ -17,8 +17,8 @@
     "dev": "npm run prisma-generate && nodemon --experimental-json-modules -e js ./src/server.js",
     "dev:queue:listen": "nodemon --experimental-json-modules -e js ./src/server/utils/service-bus-receiver.js",
     "start": "node ./src/server.js",
-    "lint:js": "eslint src prisma/seed.js",
-    "lint:js:fix": "eslint src prisma/seed.js --fix",
+    "lint:js": "eslint src prisma/seed-development.js",
+    "lint:js:fix": "eslint src prisma/seed-development.js --fix",
     "clean": "rimraf .turbo node_modules dist",
     "swagger-autogen": "node ./src/server/swagger.js",
     "db:migrate": "npx prisma migrate dev",
