+++ apps/api/prisma/seed-development.js
@@ -1,17 +1,20 @@
-import config from '../src/server/config/config.js';
 import { databaseConnector } from '../src/server/utils/database-connector.js';
 import logger from '../src/server/utils/logger.js';
+import { deleteAllRecords } from './seed-clear.js';
 import { seedStaticData } from './seed-static.js';
+import { seedTestData } from './seed-test-data.js';
 
 /**
- * Seed the production database with the required static data
+ * Clear the dev database, then add in the static and test data
  *
  * @throws {Error} If any database operation fails.
  * @returns {Promise<void>}
  */
-const seedProduction = async () => {
+const seedDevelopment = async () => {
 	try {
+		await deleteAllRecords(databaseConnector);
 		await seedStaticData(databaseConnector);
+		await seedTestData(databaseConnector);
 	} catch (error) {
 		logger.error(error);
 		throw error;
@@ -20,4 +23,4 @@ const seedProduction = async () => {
 	}
 };
 
-await (config.NODE_ENV === 'production' ? seedProduction() : null);
+await seedDevelopment();
